Move /sync/status endpoint from the find server to the ingest one
